requirements.txt
# Requirements for Automated Football Analysis Project
# Install with: pip install -r requirements.txt
#
# These match the exact libraries installed in Cell 1 of the notebook.

# ── Data access ────────────────────────────────────────────────
SoccerNet>=0.1.0          # SoccerNet dataset downloader SDK

# ── Detection model ────────────────────────────────────────────
ultralytics>=8.4.0        # YOLOv8: training, inference, ByteTrack integration

# ── Computer vision ────────────────────────────────────────────
opencv-python-headless>=4.8.0  # Frame extraction, video writing (Colab-compatible)
Pillow>=9.0.0                  # Image loading for frame size checks

# ── Deep learning (tactical CNN + tf.data pipeline) ───────────
tensorflow>=2.13.0        # tf.data dataset, Keras CNN, model training

# ── Scientific computing ───────────────────────────────────────
numpy>=1.24.0             # Array ops, heatmap generation, bbox math
matplotlib>=3.7.0         # Training curves, evaluation charts, poster figures
scikit-learn>=1.3.0       # train_test_split, class_weight, confusion matrix
pandas>=2.0.0             # StatsBomb data statistics (Cell 9)

# ── StatsBomb data (Cell 7 uses git clone, not this package)
# statsbombpy>=1.0.0      # Optional: use instead of git clone for convenience

# ── PyTorch (required by Ultralytics YOLOv8) ──────────────────
# Installed automatically with ultralytics on Colab (CUDA version)
# torch>=2.0.0
# torchvision>=0.15.0
